cli: move gitURL to api
