+++ renku/cli/_docker.py
@@ -17,97 +17,13 @@
 # limitations under the License.
 """Wrap Docker API."""
 
-import re
 import subprocess
 from configparser import NoSectionError
 
 import attr
 
 from renku import errors
-
-#: Define possible repository URLs.
-_REPOSITORY_URLS = (
-    re.compile(
-        r'^(?P<protocol>https?|git|ssh|rsync)\://'
-        r'(?:(?P<username>[^:]+)(:(?P<password>[^@]+))?@)?'
-        r'(?P<hostname>[a-z0-9_.-]*)'
-        r'[:/]*'
-        r'(?P<port>[\d]+){0,1}'
-        r'(?P<pathname>\/(?P<owner>.+)/(?P<name>.+).git)'
-    ),
-    re.compile(
-        r'(git\+)?'
-        r'((?P<protocol>\w+)://)'
-        # '((?P<user>\w+)@)?'
-        r'((?P<username>[^:]+)(:(?P<password>[^@]+))?@)?'
-        r'((?P<hostname>[\w\.\-]+))'
-        r'(:(?P<port>\d+))?'
-        r'(?P<pathname>(\/(?P<owner>\w+)/)?'
-        r'(\/?(?P<name>[\w\-]+)(\.git)?)?)'
-    ),
-    re.compile(
-        r'^(?:(?P<username>.+)@)*'
-        r'(?P<hostname>[a-z0-9_.-]*)[:/]*'
-        r'(?P<port>[\d]+){0,1}'
-        r'[:](?P<pathname>\/?(?P<owner>.+)/(?P<name>.+).git)'
-    ),
-    re.compile(
-        r'((?P<username>\w+)@)?'
-        r'((?P<hostname>[\w\.\-]+))'
-        r'[\:\/]{1,2}'
-        r'(?P<pathname>((?P<owner>\w+)/)?'
-        r'((?P<name>[\w\-]+)(\.git)?)?)'
-    ),
-    re.compile(
-        # Simple registry URL like: docker.io
-        r'((?P<hostname>[\w\.\-]+))'
-    ),
-)
-
-
-@attr.s()
-class GitURL(object):
-    """Parser for common Git URLs."""
-
-    # Initial value
-    href = attr.ib()
-    # Parsed protocols
-    pathname = attr.ib(default=None)
-    protocols = attr.ib(default=attr.Factory(list), init=False)
-    protocol = attr.ib(default='ssh')
-    hostname = attr.ib(default=None)
-    username = attr.ib(default=None)
-    password = attr.ib(default=None)
-    port = attr.ib(default=None)
-    owner = attr.ib(default=None)
-    name = attr.ib(default=None)
-
-    def __attrs_post_init__(self):
-        """Derive basic informations."""
-        if self.protocol:
-            self.protocols = self.protocol.split('+')
-
-    @classmethod
-    def parse(cls, href):
-        """Derive basic informations."""
-        for regex in _REPOSITORY_URLS:
-            if re.search(regex, href):
-                matches = re.search(regex, href)
-                return cls(href=href, **matches.groupdict())
-        else:
-            raise errors.ConfigurationError(
-                '"{href} is not a valid Git remote.'.format(href=href)
-            )
-
-    @property
-    def image(self):
-        """Return image name."""
-        img = self.hostname
-        if self.owner:
-            img += '/' + self.owner
-        if self.name:
-            img += '/' + self.name
-        return img
+from renku.api._git import GitURL
 
 
 def detect_registry_url(client, auto_login=True):
